+++ src/app/routes.tsx
@@ -1,0 +1,20 @@
+import { createBrowserRouter } from 'react-router-dom';
+import { ErrorPage } from '@pages/ErrorPage';
+import { HomePage } from '@pages/HomePage';
+import { Loader } from '@shared/ui/Loader';
+import { ItemCardDetails } from '@entities/ItemCardDetails';
+
+export const Router = createBrowserRouter([
+  {
+    path: '/',
+    element: <HomePage />,
+    errorElement: <ErrorPage />,
+    children: [
+      {
+        path: 'about_character/:characterId',
+        element: <ItemCardDetails />,
+        loader: () => <Loader />,
+      },
+    ],
+  },
+]);
