+++ src/entities/ItemCardDetails/index.ts
@@ -1,0 +1,1 @@
+export { ItemCardDetails } from './ui/ItemCardDetails';
