+++ src/entities/ItemCardDetails/ui/ItemCardDetails.tsx
@@ -1,0 +1,3 @@
+export const ItemCardDetails = () => {
+  return <div>Details here :D</div>;
+};
