+++ src/pages/ErrorPage/index.ts
@@ -1,0 +1,1 @@
+export { ErrorPage } from './ui/ErrorPage';
