+++ src/pages/ErrorPage/ui/style.module.css
@@ -1,0 +1,9 @@
+.errorWrap {
+  text-align: center;
+}
+
+.title {
+  color: rgb(0, 249, 186);
+  font-weight: bold;
+  font-size: 24px;
+}
